tools: 'brainiac' prototype, an MCP server
The idea here is to provide MCP tooling to help Claude or other things
use the repository: invoke the build system, version control, etc.

It currently is designed in terms of MCP resources and tools, but all
resources are automatically turned into tools at the moment; Claude does
not support MCP resources yet. The transformation is automatic, though.

Signed-off-by: Austin Seipp <aseipp@pobox.com>