github: always include `tilde//...` and test `//buck/tests/...`
Even when doing target determination, we want to run all the `//buck/tests/...`
tests. This is because they test all the files in the repository for problems,
but they aren't actually specified as sources to the rules. This is technically
a hack, but I'll explore a better solution in the future.

Signed-off-by: Austin Seipp <aseipp@pobox.com>