chore: migrate to watchman for development
The built in inotify watcher is at this point pretty much completely
broken; it does not respect regular expression matching on incoming file
events, and it turns out it doesn't even properly handle invalidation of
the build cache when files change (see buck2 PR 1149).

The only reason I ever defaulted to it was "it seems nice to not require
extra software", but at this point it's just far too broken for me to
justify fixing weird issues over.

Signed-off-by: Austin Seipp <aseipp@pobox.com>