rtlcc: add ulx3s-picosoc sample for testing
This is basically a miniature SOC using picorv32, designed to target the
ECP5-based ULX3S from Radiona.

The code is a bit old, but I believe originally was cobbled together
from some of Claire's original "picosoc" demos into the ULX3S "FPGA
Odysseus" tutorial code, here:

    https://github.com/ulx3s/fpga-odysseus/tree/master/tutorials/08-RISCV

I've used this example many times to test synthesis flows with Yosys as
it's relatively complete as an example, and not too long of a test. So,
I might as well commit this rather than leave it to wallow in a tmpdir
on my desktop machine...

Signed-off-by: Austin Seipp <aseipp@pobox.com>